format code, improve performance
</subject>

Branch: issue3656

<type>:
- [ ] Bug fix
- [ ] Bug fix (Test)
- [ ] New feature
- [ ] Breaking change
- [ ] Documentation update
- [ ] This change requires a documentation update

<body>

<footer>

+++ src/main/java/com/alibaba/fastjson/parser/deserializer/ASMDeserializerFactory.java
@@ -1816,7 +1816,7 @@ public class ASMDeserializerFactory implements Opcodes {
                 variantIndex += increment;
             }
             i = variants.get(name);
-            return i.intValue();
+            return i;
         }
 
         public int var(String name) {
@@ -1825,7 +1825,7 @@ public class ASMDeserializerFactory implements Opcodes {
                 variants.put(name, variantIndex++);
             }
             i = variants.get(name);
-            return i.intValue();
+            return i;
         }
 
         public int var_asm(FieldInfo fieldInfo) {
